Renderer code is back, but with another plans
