+++ graphics/renderer.c
@@ -1,0 +1,107 @@
+/*
+ * Copyright (C) 2025 Matheus Klein Schaefer
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, version 3.
+ *
+ * This program is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <https://www.gnu.org/licenses/>.
+ */
+
+/**
+ * @file shaders.c
+ * @brief Implementation file for rendering stuff
+ *
+ * @author
+ * - Matheus Klein Schaefer (email here)
+*/
+#include <SDL3/SDL.h>
+#include <graphics.h>
+
+bool Graphics_CreateAndUploadStorageBuffer(StorageBuffer *buffer,
+									void *data, size_t size)
+{
+	if(buffer == NULL || data == NULL || size <= 0)
+	{
+		SDL_LogError(SDL_LOG_CATEGORY_ERROR, "Graphics: Error: Can't generate storage buffer.");
+		return false;
+	}
+
+	buffer = SDL_CreateGPUBuffer(
+		context.device,
+		&(SDL_GPUBufferCreateInfo) {
+			.usage = SDL_GPU_BUFFERUSAGE_GRAPHICS_STORAGE_READ,
+			.size = size
+		}
+	);
+	SDL_GPUTransferBuffer* transferbuffer = SDL_CreateGPUTransferBuffer(
+		context.device,
+		&(SDL_GPUTransferBufferCreateInfo) {
+			.usage = SDL_GPU_TRANSFERBUFFERUSAGE_UPLOAD,
+			.size = size
+		}
+	);
+	void* buffer_transferdata = SDL_MapGPUTransferBuffer(context.device, transferbuffer, false);
+	SDL_memcpy(buffer_transferdata, data, size);
+	SDL_UnmapGPUTransferBuffer(context.device, transferbuffer);
+
+	SDL_GPUCommandBuffer* cmdbuf = SDL_AcquireGPUCommandBuffer(context.device);
+	SDL_GPUCopyPass* copyPass = SDL_BeginGPUCopyPass(cmdbuf);
+
+	SDL_UploadToGPUBuffer(
+		copyPass,
+		&(SDL_GPUTransferBufferLocation) {
+			.transfer_buffer = transferbuffer,
+			.offset = 0
+		},
+		&(SDL_GPUBufferRegion) {
+			.buffer = buffer,
+			.offset = 0,
+			.size = size
+		},
+		false
+	);
+
+	SDL_EndGPUCopyPass(copyPass);
+	SDL_ReleaseGPUTransferBuffer(context.device, transferbuffer);
+	SDL_SubmitGPUCommandBuffer(cmdbuf);
+
+	return true;
+}
+
+void Graphics_ReleaseStorageBuffer(StorageBuffer *buffer)
+{
+	if(buffer == NULL)
+	{
+		SDL_LogError(SDL_LOG_CATEGORY_ERROR, "Graphics: Error: Can't release invalid storage buffer.");
+		return;
+	}
+	SDL_ReleaseGPUBuffer(context.device, buffer);
+}
+
+void Graphics_SetupCommandBuffer(CommandBuffer *cmdbuf)
+{
+	cmdbuf = SDL_AcquireGPUCommandBuffer(context.device);
+}
+
+void Graphics_CommitCommandBuffer(CommandBuffer *cmdbuf)
+{
+	if(cmdbuf == NULL) return; //TODO error message
+	SDL_SubmitGPUCommandBuffer(cmdbuf);
+}
+
+void Graphics_AcquireSwapchainTexture(GPUTexture *swapchain,
+										CommandBuffer *cmdbuf)
+{
+	if(!SDL_WaitAndAcquireGPUSwapchainTexture(cmdbuf, context.window, &swapchain, NULL, NULL))
+	{
+		SDL_LogError(SDL_LOG_CATEGORY_ERROR, "Splash screen: failed to acquire swapchain texture: %s", SDL_GetError());
+		return;
+	}
+}
